// src/app/Models/ModularRequirementFormData.ts
export class ModularRequirementFormData{
    id:number;
    item_code: String;
    item_type: String;
    item_name: String;
    item_description: String;
    unit: number;
    rate: number;
    amount: number;
    piece_quantity: number;
    breadth: number;
    length: number;
    height: number;
    main_rate: number;
    fabric_m: number;
    fabric_rate: number;
    drawer_count: number;
    drawer_rate: number;
    sliding: number;
    sliding_rate:number;
    width: String
}